# Conflicts:
#	src/main/java/com/dao/UserDAO.java

+++ src/test/resources/init_dish_order.sql
@@ -1,9 +1,9 @@
 DROP TABLE IF EXISTS Dishes_Orders;
 CREATE TABLE Dishes_Orders
 (
-  id       SMALLINT(6) NOT NULL AUTO_INCREMENT PRIMARY KEY,
-  order_id SMALLINT(6) NOT NULL,
-  dish_id  SMALLINT(6) NOT NULL,
+  id       BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+  order_id BIGINT NOT NULL,
+  dish_id  BIGINT NOT NULL,
   amount   TINYINT(4)  NOT NULL,
   item_sum INT(11)     NOT NULL
 );
